Fix #46, generate better errors for Command var_types (#49)
We now allow traceback.extract_stack() to be used as well as sys.exc_info()[2]
for the tb argument to UserError. This allows stack extraction early for your
object, and a delayed rendering of the error message. Of course the other part
of the diff catches the OSError for a command not found, and presents that more
neatly (with the invoking location).